equal refactor
